+++ prefoil/utils/__init__.py
@@ -1,0 +1,2 @@
+from .io import readCoordFile, Error
+from .geom_ops import checkCellRatio
